requirements.txt
beautifulsoup4==4.12.3
Pillow==11.0.0
playwright==1.47.0
python-dotenv==1.0.1
Requests==2.32.3
